rust: use nickel-helper workspace member in build.rs
