+++ Cargo.toml
@@ -1,3 +1,9 @@
+[workspace]
+
+members = [
+    "smart-keymap-nickel-helper",
+]
+
 [package]
 name = "smart-keymap"
 version = "0.1.0"
@@ -29,6 +35,9 @@ cucumber = "0.21"
 ron = "0.8"
 serde_json = "1.0"
 futures = "0.3"
+
+[build-dependencies]
+smart-keymap-nickel-helper = { path = "./smart-keymap-nickel-helper" }
 
 [profile.dev]
 panic = "abort"
