+++ smart-keymap-nickel-helper/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "smart-keymap-nickel-helper"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
